Refactor: 외부 소통용 더미 클래스 경로 변경
