--- src/main/java/org/guzzing/studayserver/domain/like/service/AcademyAccessService.java
@@ -1,5 +1,0 @@
-package org.guzzing.studayserver.domain.like.service;
-
-public interface AcademyAccessService {
-
-}
