+++ src/main/java/org/guzzing/studayserver/domain/like/service/external/MemberAccessService.java
@@ -1,0 +1,5 @@
+package org.guzzing.studayserver.domain.like.service.external;
+
+public interface MemberAccessService {
+
+}
