+++ src/main/java/org/guzzing/studayserver/domain/like/service/external/MemberAccessServiceImpl.java
@@ -1,4 +1,4 @@
-package org.guzzing.studayserver.domain.like.service;
+package org.guzzing.studayserver.domain.like.service.external;
 
 import org.springframework.stereotype.Service;
